+++ src/Mesh.hpp
@@ -5,7 +5,7 @@
 #include "device_launch_parameters.h"
 
 extern size_t STEPS;
-extern size_t MESH_SIZE;
+extern long long int MESH_SIZE;
 
 class Mesh {
 public:
@@ -17,8 +17,10 @@ public:
 
 	void resize(size_t size) { temperature = ContainerType(size, std::vector<float>(size, INITIAL_TEMP)); }
 
-	static float getTemperature(int x, int y) {
-		if (x < 0 || y < 0 || x >= temperature[0].size() || y >= temperature.size()) { return ENVIRONMENT_TEMP; }
+	static float getTemperature(long long int x, long long int y) {
+		if (x < 0 || y < 0 || x >= static_cast<long long int>(temperature[0].size()) || y >= static_cast<long long int>(temperature.size())) {
+			return ENVIRONMENT_TEMP;
+		}
 		return temperature[y][x];
 	}
 };
